Add support for all abstract types in unboxing and throw appropriately

+++ src/demo.cpp
@@ -12,7 +12,7 @@ int main()
         end    
     )");
 
-    double res{jl("Mod.f", 2.)};
+    float res{jl("Mod.f", 2.)};
     std::printf("%f\n", res);
 
     jl_atexit_hook(0);
